Move cli.Configdir away (in `flags`).
This makes integration not depend anymore of `cli` and thus not
require `cobra` and other packages to compile.
